fix catkin, separate plane_seg_ros app and lib
