+++ plane_seg/CMakeLists.txt
@@ -63,5 +63,8 @@ install(TARGETS plane_seg
         LIBRARY DESTINATION ${CATKIN_PACKAGE_LIB_DESTINATION}
         RUNTIME DESTINATION ${CATKIN_GLOBAL_BIN_DESTINATION})
 
-# Mark header files for installation
-install(DIRECTORY include/${PROJECT_NAME}/ DESTINATION ${CATKIN_PACKAGE_INCLUDE_DESTINATION})
+install(DIRECTORY include/${PROJECT_NAME}/
+        DESTINATION ${CATKIN_PACKAGE_INCLUDE_DESTINATION})
+
+install(TARGETS block_fitter plane_seg_test plane_seg_test2
+        RUNTIME DESTINATION ${CATKIN_PACKAGE_BIN_DESTINATION})
