+++ plane_seg_ros/include/plane_seg_ros/Pass.hpp
@@ -1,0 +1,46 @@
+#pragma once
+#include <Eigen/Dense>
+#include <vector>
+#include <ros/subscriber.h>
+#include <ros/publisher.h>
+#include <ros/node_handle.h>
+#include <plane_seg/BlockFitter.hpp>
+#include <grid_map_msgs/GridMap.h>
+#include <sensor_msgs/PointCloud2.h>
+#include <geometry_msgs/PoseWithCovarianceStamped.h>
+
+namespace planeseg {
+
+class Pass{
+  public:
+    Pass(ros::NodeHandle node_);
+
+    ~Pass(){
+    }
+
+    void elevationMapCallback(const grid_map_msgs::GridMap& msg);
+    void pointCloudCallback(const sensor_msgs::PointCloud2::ConstPtr &msg);
+    void robotPoseCallBack(const geometry_msgs::PoseWithCovarianceStampedConstPtr &msg);
+
+    void processCloud(planeseg::LabeledCloud::Ptr& inCloud, Eigen::Vector3f origin, Eigen::Vector3f lookDir);
+    void processFromFile(int test_example);
+
+    void publishHullsAsCloud(std::vector<pcl::PointCloud<pcl::PointXYZ>::Ptr> cloud_ptrs,
+                                 int secs, int nsecs);
+
+    void publishHullsAsMarkers(std::vector< pcl::PointCloud<pcl::PointXYZ>::Ptr > cloud_ptrs,
+                                 int secs, int nsecs);
+    void printResultAsJson();
+    void publishResult();
+
+  private:
+    ros::NodeHandle node_;
+    std::vector<double> colors_;
+
+    ros::Subscriber point_cloud_sub_, grid_map_sub_, pose_sub_;
+    ros::Publisher received_cloud_pub_, hull_cloud_pub_, hull_markers_pub_, look_pose_pub_;
+
+    Eigen::Isometry3d last_robot_pose_;
+    planeseg::BlockFitter::Result result_;
+};
+}
