+++ plane_seg_ros/include/plane_seg_ros/geometry_utils.hpp
@@ -1,0 +1,12 @@
+#pragma once 
+#include <string>
+#include <Eigen/Dense>
+
+namespace planeseg{
+
+std::string vecToStr(const Eigen::Vector3f& iVec);
+std::string rotToStr(const Eigen::Matrix3f& iRot);
+void quat_to_euler(const Eigen::Quaterniond& q, double& roll, double& pitch, double& yaw);
+Eigen::Vector3f convertRobotPoseToSensorLookDir(Eigen::Isometry3d robot_pose);
+
+}
