+++ plane_seg_ros/src/geometry_utils.cpp
@@ -1,0 +1,41 @@
+#include "plane_seg_ros/geometry_utils.hpp"
+namespace planeseg {
+// convenience methods
+std::string vecToStr(const Eigen::Vector3f& iVec) {
+  std::ostringstream oss;
+  oss << iVec[0] << ", " << iVec[1] << ", " << iVec[2];
+  return oss.str();
+}
+
+std::string rotToStr(const Eigen::Matrix3f& iRot) {
+  std::ostringstream oss;
+  Eigen::Quaternionf q(iRot);
+  oss << q.w() << ", " << q.x() << ", " << q.y() << ", " << q.z();
+  return oss.str();
+}
+
+void quat_to_euler(const Eigen::Quaterniond& q, double& roll, double& pitch, double& yaw) {
+  const double q0 = q.w();
+  const double q1 = q.x();
+  const double q2 = q.y();
+  const double q3 = q.z();
+  roll = atan2(2.0*(q0*q1+q2*q3), 1.0-2.0*(q1*q1+q2*q2));
+  pitch = asin(2.0*(q0*q2-q3*q1));
+  yaw = atan2(2.0*(q0*q3+q1*q2), 1.0-2.0*(q2*q2+q3*q3));
+}
+
+Eigen::Vector3f convertRobotPoseToSensorLookDir(Eigen::Isometry3d robot_pose){
+
+  Eigen::Quaterniond quat = Eigen::Quaterniond( robot_pose.rotation() );
+  double r,p,y;
+  quat_to_euler(quat, r, p, y);
+  //std::cout << r*180/M_PI << ", " << p*180/M_PI << ", " << y*180/M_PI << " rpy in Degrees\n";
+
+  double yaw = y;
+  double pitch = -p;
+  double xDir = cos(yaw)*cos(pitch);
+  double yDir = sin(yaw)*cos(pitch);
+  double zDir = sin(pitch);
+  return Eigen::Vector3f(xDir, yDir, zDir);
+}
+}
