+++ plane_seg_ros/src/plane_seg_ros.cpp
@@ -1,72 +1,23 @@
+#include "plane_seg_ros/Pass.hpp"
+#include "plane_seg_ros/geometry_utils.hpp"
+
 #include <unistd.h>
 #include <ros/ros.h>
-#include <ros/console.h>
-#include <ros/package.h>
 
+#include <ros/package.h>
 
 #include <eigen_conversions/eigen_msg.h>
 #include <pcl_conversions/pcl_conversions.h>
 #include <pcl/io/pcd_io.h>
 #include <pcl/io/ply_io.h>
 
-#include <geometry_msgs/PoseStamped.h>
-#include <geometry_msgs/PoseWithCovarianceStamped.h>
 #include <visualization_msgs/Marker.h>
-#include <sensor_msgs/PointCloud2.h>
 
-#include <grid_map_msgs/GridMap.h>
 #include <grid_map_ros/grid_map_ros.hpp>
 #include <grid_map_ros/GridMapRosConverter.hpp>
+#include <geometry_msgs/PoseStamped.h>
 
-#include "plane_seg/BlockFitter.hpp"
-
-
-// convenience methods
-auto vecToStr = [](const Eigen::Vector3f& iVec) {
-  std::ostringstream oss;
-  oss << iVec[0] << ", " << iVec[1] << ", " << iVec[2];
-  return oss.str();
-};
-auto rotToStr = [](const Eigen::Matrix3f& iRot) {
-  std::ostringstream oss;
-  Eigen::Quaternionf q(iRot);
-  oss << q.w() << ", " << q.x() << ", " << q.y() << ", " << q.z();
-  return oss.str();
-};
-
-
-class Pass{
-  public:
-    Pass(ros::NodeHandle node_);
-    
-    ~Pass(){
-    }
-
-    void elevationMapCallback(const grid_map_msgs::GridMap& msg);
-    void pointCloudCallback(const sensor_msgs::PointCloud2::ConstPtr &msg);
-    void robotPoseCallBack(const geometry_msgs::PoseWithCovarianceStampedConstPtr &msg);
-
-    void processCloud(planeseg::LabeledCloud::Ptr& inCloud, Eigen::Vector3f origin, Eigen::Vector3f lookDir);
-    void processFromFile(int test_example);
-
-    void publishHullsAsCloud(std::vector<pcl::PointCloud<pcl::PointXYZ>::Ptr> cloud_ptrs,
-                                 int secs, int nsecs);
-
-    void publishHullsAsMarkers(std::vector< pcl::PointCloud<pcl::PointXYZ>::Ptr > cloud_ptrs,
-                                 int secs, int nsecs);
-    void printResultAsJson();
-    void publishResult();
-
-  private:
-    ros::NodeHandle node_;
-    std::vector<double> colors_;
-
-    ros::Subscriber point_cloud_sub_, grid_map_sub_, pose_sub_;
-    ros::Publisher received_cloud_pub_, hull_cloud_pub_, hull_markers_pub_, look_pose_pub_;
-
-    Eigen::Isometry3d last_robot_pose_;
-    planeseg::BlockFitter::Result result_;
-};
+using namespace planeseg;
 
 Pass::Pass(ros::NodeHandle node_):
     node_(node_){
@@ -124,33 +75,6 @@ void Pass::robotPoseCallBack(const geometry_msgs::PoseWithCovarianceStampedConst
   //std::cout << "got pose\n";
   tf::poseMsgToEigen(msg->pose.pose, last_robot_pose_);
 }
-
-
-void quat_to_euler(const Eigen::Quaterniond& q, double& roll, double& pitch, double& yaw) {
-  const double q0 = q.w();
-  const double q1 = q.x();
-  const double q2 = q.y();
-  const double q3 = q.z();
-  roll = atan2(2.0*(q0*q1+q2*q3), 1.0-2.0*(q1*q1+q2*q2));
-  pitch = asin(2.0*(q0*q2-q3*q1));
-  yaw = atan2(2.0*(q0*q3+q1*q2), 1.0-2.0*(q2*q2+q3*q3));
-}
-
-Eigen::Vector3f convertRobotPoseToSensorLookDir(Eigen::Isometry3d robot_pose){
-
-  Eigen::Quaterniond quat = Eigen::Quaterniond( robot_pose.rotation() );
-  double r,p,y;
-  quat_to_euler(quat, r, p, y);
-  //std::cout << r*180/M_PI << ", " << p*180/M_PI << ", " << y*180/M_PI << " rpy in Degrees\n";
-
-  double yaw = y;
-  double pitch = -p;
-  double xDir = cos(yaw)*cos(pitch);
-  double yDir = sin(yaw)*cos(pitch);
-  double zDir = sin(pitch);
-  return Eigen::Vector3f(xDir, yDir, zDir);
-}
-
 
 void Pass::elevationMapCallback(const grid_map_msgs::GridMap& msg){
   //std::cout << "got grid map / ev map\n";
@@ -463,41 +387,4 @@ void Pass::publishHullsAsMarkers(std::vector< pcl::PointCloud<pcl::PointXYZ>::Pt
 }
 
 
-int main( int argc, char** argv ){
-  // Turn off warning message about labels
-  // TODO: look into how labels are used
-  pcl::console::setVerbosityLevel(pcl::console::L_ALWAYS);
 
-
-  ros::init(argc, argv, "plane_seg");
-  ros::NodeHandle nh("~");
-  std::unique_ptr<Pass> app = std::make_unique<Pass>(nh);
-
-  ROS_INFO_STREAM("plane_seg ros ready");
-  ROS_INFO_STREAM("=============================");
-
-  bool run_test_program = false;
-  nh.param("/plane_seg/run_test_program", run_test_program, false); 
-  std::cout << "run_test_program: " << run_test_program << "\n";
-
-
-  // Enable this to run the test programs
-  if (run_test_program){
-    std::cout << "Running test examples\n";
-    app->processFromFile(0);
-    app->processFromFile(1);
-    app->processFromFile(2);
-    app->processFromFile(3);
-    // RACE examples don't work well
-    //app->processFromFile(4);
-    //app->processFromFile(5);
-
-    std::cout << "Finished!\n";
-    exit(-1);
-  }
-
-  ROS_INFO_STREAM("Waiting for ROS messages");
-  ros::spin();
-
-  return 1;
-}
